Solution

+++ app/shop.py
@@ -23,19 +23,17 @@ class Shop:
         )
 
     def print_receipt(
-            self,
-            customer_name: str,
-            product_cart: Dict[str, int]
+            self, customer_name: str, product_cart: Dict[str, int]
     ) -> None:
         print(
             f'\nDate: {datetime.datetime.now().strftime("%d/%m/%Y %H:%M:%S")}'
         )
         print(f"Thanks, {customer_name}, for your purchase!")
         print("You have bought:")
-        total = 0
+        total = 0.0
         for product, qty in product_cart.items():
             price = self.products[product]
-            cost = price * qty
+            cost = float(price * qty)
             print(f"{qty} {product}s for {format_price(cost)} dollars")
             total += cost
         print(f"Total cost is {format_price(total)} dollars")
